переделал миграции и пофиксил типы данных в контроллере

+++ I-am-Hero-API/Controllers/UsersController.cs
@@ -33,7 +33,7 @@ namespace I_am_Hero_API.Controllers
 
         // GET: api/Users/5
         [HttpGet("{id}")]
-        public async Task<ActionResult<User>> GetUser(int id)
+        public async Task<ActionResult<User>> GetUser(long id)
         {
             var user = await _context.Users.FindAsync(id);
 
@@ -48,7 +48,7 @@ namespace I_am_Hero_API.Controllers
         // PUT: api/Users/5
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
         [HttpPut("{id}")]
-        public async Task<IActionResult> PutUser(int id, User user)
+        public async Task<IActionResult> PutUser(long id, User user)
         {
             if (id != user.Id)
             {
@@ -140,7 +140,7 @@ namespace I_am_Hero_API.Controllers
 
         // DELETE: api/Users/5
         [HttpDelete("{id}")]
-        public async Task<IActionResult> DeleteUser(int id)
+        public async Task<IActionResult> DeleteUser(long id)
         {
             var user = await _context.Users.FindAsync(id);
             if (user == null)
@@ -154,7 +154,7 @@ namespace I_am_Hero_API.Controllers
             return NoContent();
         }
 
-        private bool UserExists(int id)
+        private bool UserExists(long id)
         {
             return _context.Users.Any(e => e.Id == id);
         }
